Register new snapshots
