+++ src/snapshots.txt
@@ -1,3 +1,11 @@
+S 2013-10-29 fed48cc
+  freebsd-x86_64 4a43216b432511a5fd6b727753aedb749f6a68dc
+  linux-i386 53f65d4b1377c17fc12d05d7c4a0fbd92eea071f
+  linux-x86_64 afa5f19a37a2cf137e5d4277951fa07efda8e072
+  macos-i386 7522c24f78ed35020e2877e3eada058ea8a11f35
+  macos-x86_64 a18afdcbbdbb81c1fdf08788b24f0d3ea8701eb1
+  winnt-i386 c78f0839c9524eda33c54a5232121886021b5352
+
 S 2013-10-28 2ab4a6f
   freebsd-x86_64 08af04bcf739930bdb7d0ad244b2c8094cd5096a
   linux-i386 c233de1ed09872d5c7a3e1ce9ab9eb6e16631201
